refactor(models): share catalog scroll sentinel
