+++ desktop/src/mobile/MobileCatalogView.vue
@@ -42,6 +42,7 @@ import {
 } from "@studio/lib/modelInstallTargets";
 import { catalogThumbnailUrl } from "../lib/catalogThumbnails";
 import { isVideoFamily } from "../lib/capabilities";
+import { useInfiniteScrollSentinel } from "../lib/useInfiniteScrollSentinel";
 import { formatCount, formatGB, percent } from "../lib/format";
 import {
   isCatalogModelId,
@@ -640,44 +641,8 @@ function loadMore(): void {
   void runSearch(false);
 }
 
-/** End-of-list sentinel: scrolling near it fetches the next page — the
- *  old Load more button read as a dead end and (worse) looked broken
- *  whenever upstream paging returned duplicate rows the dedup swallowed. */
 const sentinel = ref<HTMLElement | null>(null);
-const sentinelVisible = ref(false);
-/** Fetches chained since the last real intersection event — bounded like
- *  the media-filter auto-fetch so a dedup-swallowed page can't loop. */
-let chainedFetches = 0;
-let sentinelObserver: IntersectionObserver | null = null;
-
-watch(sentinel, (el) => {
-  sentinelObserver?.disconnect();
-  sentinelObserver = null;
-  sentinelVisible.value = false;
-  if (!el) return;
-  sentinelObserver = new IntersectionObserver(
-    (hits) => {
-      for (const hit of hits) sentinelVisible.value = hit.isIntersecting;
-      if (sentinelVisible.value) {
-        chainedFetches = 0;
-        loadMore();
-      }
-    },
-    { rootMargin: "600px 0px" },
-  );
-  sentinelObserver.observe(el);
-});
-
-// A page that lands without pushing the sentinel out of view (dedup or the
-// media filter swallowed its rows) emits no new intersection event, so the
-// observer alone would stall. Chain the next fetch while the sentinel is
-// still visible, up to the auto-fetch bound.
-watch(loading, (isLoading) => {
-  if (isLoading || !sentinelVisible.value || !hasMore.value) return;
-  if (chainedFetches >= MAX_AUTO_PAGES) return;
-  chainedFetches += 1;
-  loadMore();
-});
+useInfiniteScrollSentinel(sentinel, loading, hasMore, loadMore, MAX_AUTO_PAGES);
 
 async function loadFamilies(): Promise<void> {
   const target = selectedTarget.value;
@@ -1132,7 +1097,6 @@ onBeforeUnmount(() => {
   ++modelsEpoch;
   ++detailEpoch;
   if (debounce) clearTimeout(debounce);
-  sentinelObserver?.disconnect();
   deactivateInteractions();
   mobileDownloads.unregisterConsumer(DOWNLOAD_CONSUMER_ID);
 });
